[DELTADNA-758] - Fix changing language bug
A user reported crashes when changing their OS language from English to Turkish.

The bug was caused because column names in our SDK database are lowercased, which is a locale sensitive operation in Java. This meant on devices set to use a language with a non-Latin alphabet (e.g. Turkish) the database columns were inaccessible.

.detDefault() will set the locale to the default on the OS, which will be setting the language back to Turkish and causing the crash.